Refactored LedDisplay: digits are now defined as uint8_t, instead of an array of bools.

+++ MiniCalculator/Calc.h
@@ -15,7 +15,7 @@
 using opfun = long(*)(long, long);
 
 struct OperatorFun {
-	int Name[4];
+	uint8_t Name[4];
 	opfun fun;
 };
